requirements.txt
opencv-python
PyYAML
torch
numpy
scipy
face-alignment